feat: add orthographic projection
