+++ Cargo.toml
@@ -7,12 +7,16 @@ edition = "2024"
 boolygon = { git = "https://github.com/hectormrc/boolygon.git", default-features = false }
 geocart = { git = "https://github.com/hectormrc/geocart.git", default-features = false }
 num-traits = { version = "0.2.19", default-features = false }
-serde = { version = "1.0.219", default-features = false, features = ["std", "derive"], optional = true}
+serde = { version = "1.0.219", default-features = false, features = [
+  "std",
+  "derive",
+], optional = true }
 thiserror = { version = "2.0.12", default-features = false }
 
 [features]
-default = ["equirectangular", "gall-stereographic", "orthographic", "serde"]
+default = ["equirectangular", "gall-stereographic", "orthographic", "serde", "svg"]
 equirectangular = []
 gall-stereographic = []
 orthographic = []
-serde = ["dep:serde"]
+serde = ["dep:serde", "geocart/serde"]
+svg = []
