+++ src/svg/ser.rs
@@ -1,0 +1,8 @@
+use serde::ser::{self, Impossible, Serialize};
+
+/// A SVG serializer.
+pub struct Serializer<W> {
+    writter: W,
+}
+
+//impl<W> serde::ser::Serializer for Serializer<W> {}
